+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - added a new callback: `ToggleToolbarButtonCallback`
 - changed behavior of Search: results are kept in memory while the window is not manually closed
 - added a new `AbstractWorkspaceItemCreator` component, to wrap components used to create MyWorkspace items.
+- updated the style of the `VcsExpansionPanel` and `VcsGroupedList` compononents to harmonize them with the `VcsTreeview` component
 
 # 6.1.14